+++ CHANGELOG.md
@@ -21,6 +21,7 @@ Unless otherwise specified, any version comparison below is the comparison of th
 - Added support for loading and sending behavior packs (NetEase client requires sending behavior packs to the client).
 - Implemented tripwire hook and tripwire.
 - Implemented armor stand.
+- Improved entity tasks handling. Now the Entity Manager task queue uses the same handling strategy as the synced packets, and the entity will be spawned/removed as soon as possible, rather than waiting for the next game tick.
 
 ### Changed